Remove obsolete cast to size_t
drizzle_column_max_size already returns size_t

+++ tests/unit/column.c
@@ -122,7 +122,7 @@ int main(int argc, char *argv[])
       case 2:
         ASSERT_EQ_(drizzle_column_max_size(column), 255,
                    "Column max size wrong %" PRIu64 " != 255",
-                   (size_t)drizzle_column_max_size(column));
+                   drizzle_column_max_size(column));
 
         ASSERT_EQ_(drizzle_column_charset(column), DRIZZLE_CHARSET_LATIN1_SWEDISH_CI,
                    "Column charset wrong, %d != %d", drizzle_column_charset(column),
